src/fake_detection_api/requirements.txt
easydict==1.9
numpy==1.17.0
tqdm==4.31.1
torchvision==0.4.0
torch==1.2.0
opencv_python==4.2.0.34
Pillow==7.1.2
tensorboardX==2.0

Flask==1.1.1
Flask-RESTful==0.3.7

# Pillow==6.2.1

sentry-sdk[flask]==0.14.3